Use the load-grunt-config plugin to split the gruntfile into a file for each task
This is better for scalability and maintainability. It allows easier updating
between bread and butter versions. Specifically a change to the copy task in
the bread and butter core can be applied to an old project by just replcing
that file, not the entire gruntfile that may have been modified for the specific
project.