Resolve #33. Exclude Modernizer from the bower_concat task.
Also add it to the uglify task and include in header.php. This will work once modernizer is added via bower. I have not installed modernizer in the bread and butter template because I want people to install it themselves and answer all the questions Bower asks in the setup.